+++ src/dataset/webqsp_sample.py
@@ -87,9 +87,9 @@ def preprocess(sample_size: int, seed: int, retrieval_method: str, tele_mode: st
 
 def sample_dataset(dataset, sample_size: int, seed: int):
     np.random.seed(seed)
-    train_size = min(sample_size * 2.4, len(dataset['train']))
-    val_size = min(sample_size * 0.8, len(dataset['validation']))
-    test_size = min(sample_size * 0.8, len(dataset['test']))
+    train_size = min(int(sample_size * 2.4), len(dataset['train']))
+    val_size = min(int(sample_size * 0.8), len(dataset['validation']))
+    test_size = min(int(sample_size * 0.8), len(dataset['test']))
 
     train_indices = np.random.choice(len(dataset['train']), size=train_size, replace=False)
     val_indices = np.random.choice(len(dataset['validation']), size=val_size, replace=False)
